app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
psycopg2-binary==2.9.9
llama-index==0.10.0
voyageai==0.1.0
pydantic==2.6.1
